Change new user check to use getUserData
`authplain`'s `modifyUser` function outputs an error message via the `msg` function when called for a user that does not exist. Due to other issues with `getLang` and scoping, this just creates an error box with no message, as no lang mappings exist for this plugin that match those used in `authplain`'s code. An error message would not make sense anyways. I think it makes more sense to use `getUserData` and check for a falsey value.